requirements.txt
geopandas==0.14.0
matplotlib==3.8
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==10.1.0
rasterio==1.3.9
scikit_image==0.22.0
scipy==1.11.3
Shapely==2.0.2
